Base grid implementation using paramQueryGrid (based on plain jQueryUI). Typings definition for jQueryUI has been enriched with pqGrid interfaces.
BaseGrid is control to represent grid data for specific entity (IDataStructure).
It has replaced EntityListComponent duties.
EntityListComponent becomes obsolete, but still can be used as example for full custom representation of list of entities.
For RestaurantListComponent override is used (customization), so that base grid is hidden and full custom EntityList is used.